chore: refactor on sdk client packages
