+++ pkg/sdkclient/serverinfo/serverinfo_test.go
@@ -1,4 +1,4 @@
-package util
+package serverinfo
 
 import (
 	"testing"
