+++ pkg/sdkclient/sessionreducer/client.go
@@ -21,13 +21,14 @@ import (
 	"errors"
 	"io"
 
-	sessionreducepb "github.com/numaproj/numaflow-go/pkg/apis/proto/sessionreduce/v1"
-	"github.com/numaproj/numaflow-go/pkg/info"
 	"google.golang.org/grpc"
 	"google.golang.org/protobuf/types/known/emptypb"
 
+	sessionreducepb "github.com/numaproj/numaflow-go/pkg/apis/proto/sessionreduce/v1"
+	"github.com/numaproj/numaflow-go/pkg/info"
 	"github.com/numaproj/numaflow/pkg/sdkclient"
-	"github.com/numaproj/numaflow/pkg/shared/util"
+	sdkerr "github.com/numaproj/numaflow/pkg/sdkclient/error"
+	grpcutil "github.com/numaproj/numaflow/pkg/sdkclient/grpc"
 )
 
 // client contains the grpc connection and the grpc client.
@@ -45,7 +46,7 @@ func New(serverInfo *info.ServerInfo, inputOptions ...sdkclient.Option) (Client,
 	}
 
 	// Connect to the server
-	conn, err := util.ConnectToServer(opts.UdsSockAddr(), serverInfo, opts.MaxMessageSize())
+	conn, err := grpcutil.ConnectToServer(opts.UdsSockAddr(), serverInfo, opts.MaxMessageSize())
 	if err != nil {
 		return nil, err
 	}
@@ -88,7 +89,7 @@ func (c *client) SessionReduceFn(ctx context.Context, datumStreamCh <-chan *sess
 
 	if err != nil {
 		go func() {
-			errCh <- util.ToUDFErr("c.grpcClt.SessionReduceFn", err)
+			errCh <- sdkerr.ToUDFErr("c.grpcClt.SessionReduceFn", err)
 		}()
 	}
 
@@ -105,7 +106,7 @@ func (c *client) SessionReduceFn(ctx context.Context, datumStreamCh <-chan *sess
 					break outerLoop
 				}
 				if sendErr = stream.Send(datum); sendErr != nil && !errors.Is(sendErr, io.EOF) {
-					errCh <- util.ToUDFErr("SessionReduceFn stream.Send()", sendErr)
+					errCh <- sdkerr.ToUDFErr("SessionReduceFn stream.Send()", sendErr)
 					return
 				}
 			}
@@ -118,7 +119,7 @@ func (c *client) SessionReduceFn(ctx context.Context, datumStreamCh <-chan *sess
 		default:
 			sendErr = stream.CloseSend()
 			if sendErr != nil && !errors.Is(sendErr, io.EOF) {
-				errCh <- util.ToUDFErr("SessionReduceFn stream.CloseSend()", sendErr)
+				errCh <- sdkerr.ToUDFErr("SessionReduceFn stream.CloseSend()", sendErr)
 			}
 		}
 	}()
@@ -143,7 +144,7 @@ func (c *client) SessionReduceFn(ctx context.Context, datumStreamCh <-chan *sess
 					return
 				}
 				if recvErr != nil {
-					errCh <- util.ToUDFErr("SessionReduceFn stream.Recv()", recvErr)
+					errCh <- sdkerr.ToUDFErr("SessionReduceFn stream.Recv()", recvErr)
 				}
 				responseCh <- resp
 			}
