+++ pkg/sdkclient/sourcetransformer/client.go
@@ -19,13 +19,14 @@ package sourcetransformer
 import (
 	"context"
 
-	transformpb "github.com/numaproj/numaflow-go/pkg/apis/proto/sourcetransform/v1"
-	"github.com/numaproj/numaflow-go/pkg/info"
 	"google.golang.org/grpc"
 	"google.golang.org/protobuf/types/known/emptypb"
 
+	transformpb "github.com/numaproj/numaflow-go/pkg/apis/proto/sourcetransform/v1"
+	"github.com/numaproj/numaflow-go/pkg/info"
 	"github.com/numaproj/numaflow/pkg/sdkclient"
-	"github.com/numaproj/numaflow/pkg/shared/util"
+	sdkerr "github.com/numaproj/numaflow/pkg/sdkclient/error"
+	grpcutil "github.com/numaproj/numaflow/pkg/sdkclient/grpc"
 )
 
 // client contains the grpc connection and the grpc client.
@@ -43,7 +44,7 @@ func New(serverInfo *info.ServerInfo, inputOptions ...sdkclient.Option) (Client,
 	}
 
 	// Connect to the server
-	conn, err := util.ConnectToServer(opts.UdsSockAddr(), serverInfo, opts.MaxMessageSize())
+	conn, err := grpcutil.ConnectToServer(opts.UdsSockAddr(), serverInfo, opts.MaxMessageSize())
 	if err != nil {
 		return nil, err
 	}
@@ -81,7 +82,7 @@ func (c *client) IsReady(ctx context.Context, in *emptypb.Empty) (bool, error) {
 // SourceTransformFn SourceTransformerFn applies a function to each request element.
 func (c *client) SourceTransformFn(ctx context.Context, request *transformpb.SourceTransformRequest) (*transformpb.SourceTransformResponse, error) {
 	transformResponse, err := c.grpcClt.SourceTransformFn(ctx, request)
-	err = util.ToUDFErr("c.grpcClt.SourceTransformFn", err)
+	err = sdkerr.ToUDFErr("c.grpcClt.SourceTransformFn", err)
 	if err != nil {
 		return nil, err
 	}
