+++ pkg/sideinputs/manager/manager.go
@@ -21,14 +21,14 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/numaproj/numaflow/pkg/sdkserverinfo"
-
 	cronlib "github.com/robfig/cron/v3"
 	"go.uber.org/zap"
 	"google.golang.org/protobuf/types/known/emptypb"
 
 	dfv1 "github.com/numaproj/numaflow/pkg/apis/numaflow/v1alpha1"
 	"github.com/numaproj/numaflow/pkg/isbsvc"
+	"github.com/numaproj/numaflow/pkg/sdkclient"
+	sdkserverinfo "github.com/numaproj/numaflow/pkg/sdkclient/serverinfo"
 	"github.com/numaproj/numaflow/pkg/sdkclient/sideinput"
 	jsclient "github.com/numaproj/numaflow/pkg/shared/clients/nats"
 	"github.com/numaproj/numaflow/pkg/shared/kvs"
@@ -83,7 +83,7 @@ func (sim *sideInputsManager) Start(ctx context.Context) error {
 	}
 
 	// Wait for server info to be ready
-	serverInfo, err := sdkserverinfo.SDKServerInfo(sdkserverinfo.WithServerInfoFilePath(sdkserverinfo.SideInputServerInfoFile))
+	serverInfo, err := sdkserverinfo.SDKServerInfo(sdkserverinfo.WithServerInfoFilePath(sdkclient.SideInputServerInfoFile))
 	if err != nil {
 		return err
 	}
